Workaround iOS emulators keychain problems
iOS emulators need a keychain-access-groups key in
the entitlements.plist file in order to enable the
keychain and then use the secure storage service. To be
able to use a custom entitlements.plist file you need
a provisioning profile. For getting a provisioning profile
you need an Apple Developer program enrollment. For that
you need to pay 100 euros per year.

https://docs.microsoft.com/en-us/xamarin/essentials/secure-storage?tabs=ios